Moves fixtures to loqusdb own conftest
